+++ tests/fixtures/fixtures_storage.py
@@ -39,4 +39,4 @@ def storage_mock(auth_mock, requests_mock):
     # orders info
     url_order_info = f"{constants.API_HOST}/v2/orders/{constants.ORDER_ID}"
     requests_mock.get(url=url_order_info, json=constants.JSON_ORDER)
-    return storage.Storage(auth=auth_mock)
+    return storage.Storage(auth=auth_mock, workspace_id=constants.WORKSPACE_ID)