+++ tests/test_storage.py
@@ -145,7 +145,7 @@ def test_get_assets_pagination(auth_mock, requests_mock):
     url_storage_assets_paginated = f"{constants.API_HOST}/v2/assets?sort=createdAt,asc&size=50"
     requests_mock.get(url=url_storage_assets_paginated, json=json_assets_paginated)
 
-    storage_results = storage.Storage(auth=auth_mock)
+    storage_results = storage.Storage(auth=auth_mock, workspace_id=constants.WORKSPACE_ID)
     assets = storage_results.get_assets(limit=74, sortby="createdAt", descending=False)
     assert len(assets) == 74
     assert isinstance(assets[0], asset.Asset)
@@ -272,7 +272,7 @@ def test_get_orders_v2_endpoint_params(auth_mock, requests_mock, params, expecte
     url_params = "&".join(
         [
             "sort=createdAt%2Cdesc",
-            f"workspaceId={constants.WORKSPACE_ID}" if params["workspace_orders"] else "",
+            (f"workspaceId={constants.WORKSPACE_ID}" if params["workspace_orders"] else ""),
             f"""displayName={params["name"]}""" if params["name"] else "",
             *[f"status={status}" for status in endpoint_statuses],
             "size=50",
@@ -291,7 +291,7 @@ def test_get_orders_v2_endpoint_params(auth_mock, requests_mock, params, expecte
             )
             for output in expected_results
         ]
-    storage_results = storage.Storage(auth=auth_mock)
+    storage_results = storage.Storage(auth=auth_mock, workspace_id=constants.WORKSPACE_ID)
     orders = storage_results.get_orders(**params)
     assert orders == expected_results
 
@@ -328,11 +328,11 @@ def test_get_orders_pagination(auth_mock, requests_mock):
 
     # assets pages
     url_storage_orders_paginated = (
-        f"{constants.API_HOST}/v2/orders?sort=createdAt,asc&workspaceId={auth_mock.workspace_id}&size=50"
+        f"{constants.API_HOST}/v2/orders?sort=createdAt,asc&workspaceId={constants.WORKSPACE_ID}&size=50"
     )
     requests_mock.get(url=url_storage_orders_paginated, json=json_orders_paginated)
 
-    storage_results = storage.Storage(auth=auth_mock)
+    storage_results = storage.Storage(auth=auth_mock, workspace_id=constants.WORKSPACE_ID)
     orders = storage_results.get_orders(limit=74, sortby="createdAt", descending=False)
     assert len(orders) == 74
     assert isinstance(orders[0], order.Order)