+++ up42/initialization.py
@@ -8,61 +8,55 @@ DEPRECATION_MESSAGE = "after May 15th, 2024, and will be replaced by new process
 INITIALIZED_MSG = "Initialized %s"
 
 
-@main.check_auth
 def initialize_catalog() -> catalog.Catalog:
     """
     Returns a Catalog object for using the catalog search.
     """
-    return catalog.Catalog(auth=main.get_auth_safely())
+    return catalog.Catalog(auth=main.workspace.auth, workspace_id=main.workspace.id)
 
 
-@main.check_auth
 def initialize_tasking() -> "tasking.Tasking":
     """
     Returns a Tasking object for creating satellite tasking orders.
     """
-    return tasking.Tasking(auth=main.get_auth_safely())
+    return tasking.Tasking(auth=main.workspace.auth, workspace_id=main.workspace.id)
 
 
-@main.check_auth
 def initialize_storage() -> storage.Storage:
     """
     Returns a Storage object to list orders and assets.
     """
-    return storage.Storage(auth=main.get_auth_safely())
+    return storage.Storage(auth=main.workspace.auth, workspace_id=main.workspace.id)
 
 
-@main.check_auth
 def initialize_order(order_id: str) -> order.Order:
     """
     Returns an Order object (has to exist on UP42).
     Args:
         order_id: The UP42 order_id
     """
-    up42_order = order.Order(auth=main.get_auth_safely(), order_id=order_id)
+    up42_order = order.Order(auth=main.workspace.auth, order_id=order_id)
     logger.info(INITIALIZED_MSG, up42_order)
     return up42_order
 
 
-@main.check_auth
 def initialize_asset(asset_id: str) -> asset.Asset:
     """
     Returns an Asset object (has to exist on UP42).
     Args:
         asset_id: The UP42 asset_id
     """
-    up42_asset = asset.Asset(auth=main.get_auth_safely(), asset_id=asset_id)
+    up42_asset = asset.Asset(auth=main.workspace.auth, asset_id=asset_id)
     logger.info(INITIALIZED_MSG, up42_asset)
     return up42_asset
 
 
-@main.check_auth
 def initialize_webhook(webhook_id: str) -> webhooks.Webhook:
     """
     Returns a Webhook object (has to exist on UP42).
     Args:
         webhook_id: The UP42 webhook_id
     """
-    webhook = webhooks.Webhook(auth=main.get_auth_safely(), webhook_id=webhook_id)
+    webhook = webhooks.Webhook(auth=main.workspace.auth, workspace_id=main.workspace.id, webhook_id=webhook_id)
     logger.info(INITIALIZED_MSG, webhook)
     return webhook